Added scalar type to ScalarImage class
So far scalar images were fixed to be of value type float. This
restriction complicates the code and is unnecessary. This commit changes
it to be of any arbitrary scalar type.

+++ src/test/scala/scalismo/numerics/IntegrationTest.scala
@@ -40,7 +40,7 @@ class IntegrationTest extends ScalismoTestSuite {
       val integrator = Integrator[_1D](GridSampler(grid))
 
       val res = integrator.integrateScalar(img)
-      res should be((1.0 / 3.0).toFloat +- 0.01)
+      res should be((1.0 / 3.0) +- 0.01)
     }
 
     it("Correctly integrates sin(x) on interval [-Pi, Pi]") {
